Revert "Merge branch 'TheStrategists' into SebastianVargas"
This reverts commit 31fe984ce38543fc787d0d1814d9b612e3632226, reversing
changes made to d313e7867114fe88f4a8bf03efffd651c8df8904.

+++ Opiniometro_Datos/Opiniometro_Datos/Permiso.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Permiso]
 (
-	[Id]			SMALLINT		NOT NULL,
+	[Id]			TINYINT			NOT NULL,
 	[Descripcion]	NVARCHAR(150)	NOT NULL,
 	CONSTRAINT PK_Permiso
 		PRIMARY KEY (Id)
