+++ Opiniometro_WebApp/Opiniometro_WebApp/Models/Posee_Enfasis_Perfil_Permiso.cs
@@ -17,7 +17,7 @@ namespace Opiniometro_WebApp.Models
         public byte NumeroEnfasis { get; set; }
         public string SiglaCarrera { get; set; }
         public string IdPerfil { get; set; }
-        public short IdPermiso { get; set; }
+        public byte IdPermiso { get; set; }
     
         public virtual Perfil Perfil { get; set; }
         public virtual Permiso Permiso { get; set; }
